adds EditProfile.tsx error handle
