+++ packages/client/src/pages/ProfilePage/Profile.tsx
@@ -13,7 +13,6 @@ import { useAppDispatch } from '../../hooks/redux';
 import { userState } from '../../hooks/userState';
 import { fetchProfileChange } from '../../store/User/profile/actions';
 import ProfileAvatar from './ProfileAvatar';
-import { errorReset } from '../../store/User/userSlice';
 import { TITLES, useTitle } from 'utils/useTitle';
 
 const Profile: FC = () => {
@@ -21,7 +20,7 @@ const Profile: FC = () => {
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
 
-  const { user, userError } = userState();
+  const { user } = userState();
 
   const avatar = <ProfileAvatar image={user!.avatar} />;
   const title = <StUserName>{user?.first_name}</StUserName>;
@@ -56,9 +55,6 @@ const Profile: FC = () => {
       handleChangePassword={handleChangePassword}
     />
   );
-  const errorCancel = () => {
-    if (userError) dispatch(errorReset());
-  };
 
   return (
     <StyledForm
@@ -69,8 +65,6 @@ const Profile: FC = () => {
       defaultValues={defaultValues}
       footer={footer}
       inputCss={inputCss}
-      error={userError}
-      errorReset={errorCancel}
     />
   );
 };
